Add election_book family (G3) + migrate broome (upright SOVC PDF override)
